requirements.txt
playwright==1.40.0
snownlp==0.12.3
pandas==2.1.1
datetime==5.2
python-dotenv==1.0.0
requests==2.31.0
